+++ src/app/app.component.scss
@@ -1,4 +1,4 @@
-@import "../variables.scss";
+@import "../sass/variables.scss";
 
 * {
     margin: 0;
