+++ src/app/backend/assets/constant/status.enum.ts
@@ -1,0 +1,2 @@
+export enum Status {
+}
