+++ src/app/backend/assets/templates/metadata.d.ts
