+++ src/app/components/body/queue/queue.component.scss
