+++ src/app/components/footer/main-bar/main-bar.component.scss
@@ -1,4 +1,4 @@
-@import "../../variables.scss";
+@import "../../../../sass/variables.scss";
 
 #main-bar {
     display: flex;
