+++ src/app/components/footer/main-bar/main-bar.component.spec.ts
