+++ src/app/components/footer/preset-selector/preset-selector.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { PresetService, Preset } from './preset.service';
+import { PresetService, Preset } from '../../../backend/services/preset.service';
 
 @Component({
   selector: 'app-preset-selector',
