+++ src/app/components/footer/url-bar/url-bar.component.scss
