+++ src/app/components/footer/url-bar/url-bar.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { PreviewService } from '../preview.service';
+import { PreviewService } from '../../../backend/services/preview.service';
 
 @Component({
   selector: 'app-url-bar',
