+++ src/app/components/header/header.component.scss
@@ -1,4 +1,4 @@
-@import "../../variables.scss";
+@import "../../../sass/variables.scss";
 @import url('https://fonts.googleapis.com/css2?family=Press+Start+2P&display=swap');
 
 $hover-color: rgba(255,255,255,0.1);
