+++ src/app/components/header/windowCtrl/close-winCtrl.component.ts
