+++ src/app/components/header/windowCtrl/min-winCtrl.component.ts
