+++ src/app/dialogs/PSettings/p-set-dialog-body/p-set-dialog-body.component.scss
@@ -1,4 +1,4 @@
-@import "../../variables.scss";
+@import "../../../../sass/variables.scss";
 
 app-p-set-dialog-body {
     display: block;
@@ -28,6 +28,6 @@ app-p-set-dialog-body {
     background: $primary;
 }
 
-.dialog {
+// .dialog {
     
-}
+// }
