+++ src/styles.scss
@@ -1,6 +1,6 @@
 /* You can add global styles to this file, and also import other style files */
 
-@import "./variables.scss";
+@import "./sass/variables.scss";
 
 body {
     background-color: $primary-light;
